use vue-cli-plugin-electron-builder

+++ packages/web/src/background.ts
@@ -1,33 +1,37 @@
-import { fork } from 'child_process'
 import path from 'path'
+import { fork } from 'child_process'
 
 import { app, protocol, BrowserWindow } from 'electron'
+import {
+  createProtocol,
+  installVueDevtools
+} from 'vue-cli-plugin-electron-builder/lib'
 import contextMenu from 'electron-context-menu'
 // @ts-ignore
 import ON_DEATH from 'death'
-
-import { createProtocol } from './vue-cli-plugin'
 
 const isDevelopment = process.env.NODE_ENV !== 'production'
 app.allowRendererProcessReuse = true
 
 contextMenu()
 
-const p = fork(path.join(__dirname, './server/index.js'), [], {
-  stdio: 'inherit'
-})
+if (!isDevelopment) {
+  const p = fork(path.join(__dirname, './server/index.js'), [], {
+    stdio: 'inherit'
+  })
 
-ON_DEATH(() => {
-  if (!p.killed) {
-    p.kill()
-  }
-})
+  ON_DEATH(() => {
+    if (!p.killed) {
+      p.kill()
+    }
+  })
 
-app.once('before-quit', () => {
-  if (!p.killed) {
-    p.kill()
-  }
-})
+  app.once('before-quit', () => {
+    if (!p.killed) {
+      p.kill()
+    }
+  })
+}
 
 // Keep a global reference of the window object, if you don't, the window will
 // be closed automatically when the JavaScript object is garbage collected.
@@ -48,8 +52,15 @@ function createWindow () {
   win.maximize()
 
   createProtocol('app')
-  // Load the index.html when not in development
-  win.loadURL('app://./index.html')
+  if (process.env.WEBPACK_DEV_SERVER_URL) {
+    // Load the url of the dev server if in development mode
+    win.loadURL(process.env.WEBPACK_DEV_SERVER_URL as string)
+    // if (!process.env.IS_TEST) win.webContents.openDevTools()
+  } else {
+    createProtocol('app')
+    // Load the index.html when not in development
+    win.loadURL('app://./index.html')
+  }
 
   win.on('closed', () => {
     win = null
@@ -85,7 +96,6 @@ app.on('ready', async () => {
     // If you are not using Windows 10 dark mode, you may uncomment these lines
     // In addition, if the linked issue is closed, you can upgrade electron and uncomment these lines
     try {
-      const installVueDevtools = (await import('./vue-cli-plugin/installVueDevtools')).default
       await installVueDevtools()
     } catch (e) {
       console.error('Vue Devtools failed to install:', e.toString())
